Create a new Azure Machine Learning workspace

Create a new Azure Machine Learning workspace. Throws an exception if the workspace already exists or any of the workspace requirements are not satisfied. When you create new workspace, it automatically creates several Azure resources that are used in the workspace:

Azure Container Registry: Registers Docker containers that you use during training and when you deploy a model. To minimize costs, ACR is lazy-loaded until deployment images are created.

Azure Storage account: Used as the default datastore for the workspace.

Azure Application Insights: Stores monitoring information about your models.

Azure Key Vault: Stores secrets that are used by compute targets and other sensitive information that's needed by the workspace.

create_workspace(name, subscription_id = NULL, resource_group = NULL,
  location = NULL, create_resource_group = TRUE,
  friendly_name = NULL, storage_account = NULL, key_vault = NULL,
  app_insights = NULL, container_registry = NULL, exist_ok = FALSE,
  show_output = TRUE)

Arguments

name
A string of the new workspace name. Workspace name has to be between 2 and 32 characters of letters and numbers.

subscription_id
A string of the subscription ID of the containing subscription for the new workspace. The parameter is required if the user has access to more than one subscription.

resource_group
A string of the Azure resource group that is containing the workspace. The parameter defaults to a mutation of the workspace name.

location
A string of the location of the workspace. The parameter defaults to the resource group location. The location has to be a supported region for Azure Machine Learning Services.

create_resource_group
If TRUE the resource group will be created if it doesn't exist.

friendly_name
A string of the friendly name for the workspace that can be displayed in the UI.

storage_account
A string of an existing storage account in the Azure resource ID format. The storage will be used by the workspace to save run outputs, code, logs etc. If NULL a new storage will be created.

key_vault
A string of an existing key vault in the Azure resource ID format. The key vault will be used by the workspace to store credentials added to the workspace by the users. If NULL a new key vault will be created.

app_insights
A string of an existing Application Insights in the Azure resource ID format. The Application Insights will be used by the workspace to log webservices events. If NULL a new Application Insights will be created.

container_registry
A string of an existing container registry in the Azure resource ID format. The container registry will be used by the workspace to pull and push both experimentation and webservices images. If NULL a new container registry will be created.

exist_ok
If TRUE the method will not fail if the workspace already exists.

show_output
If TRUE the method will print out incremental progress of method.

Value

The Workspace object.

Examples

# This example requires only minimal specification, and all dependent
# resources as well as the resource group will be created automatically.
if (FALSE) {
ws <- create_workspace(name = 'myworkspace',
                       subscription_id = '<azure-subscription-id>',
                       resource_group = 'myresourcegroup',
                       location = 'eastus2')
}

# This example shows how to reuse existing Azure resources by making
# use of all parameters utilizing the Azure resource ID format. The specific
# Azure resource IDs can be retrieved through the Azure Portal or SDK. This
# assumes that the resource group, storage account, key vault, App Insights
# and container registry already exist
if (FALSE) {
ws <- create_workspace(
       name = 'myworkspace',
       subscription_id = '<azure-subscription-id>',
       resource_group = 'myresourcegroup',
       create_resource_group = FALSE,
       location = 'eastus2',
       friendly_name = 'My workspace',
       storage_account = 'subscriptions/<azure-subscription-id>/resourcegroups/myresourcegroup/providers/microsoft.storage/storageaccounts/mystorageaccount',
       key_vault = 'subscriptions/<azure-subscription-id>/resourcegroups/myresourcegroup/providers/microsoft.keyvault/vaults/mykeyvault',
       app_insights = 'subscriptions/<azure-subscription-id>/resourcegroups/myresourcegroup/providers/microsoft.insights/components/myappinsights',
       container_registry = 'subscriptions/<azure-subscription-id>/resourcegroups/myresourcegroup/providers/microsoft.containerregistry/registries/mycontainerregistry')
}