Represents a path to data in a datastore.

The path represented by DataPath object can point to a directory or a data artifact (blob, file).

data_path(datastore, path_on_datastore = NULL, name = NULL)

Arguments

datastore
The Datastore to reference.

path_on_datastore
The relative path in the backing storage for the data reference.

name
An optional name for the DataPath.

Value

The DataPath object.

Examples

my_data <- register_azure_blob_container_datastore(
    workspace = ws,
    datastore_name = blob_datastore_name,
    container_name = ws_blob_datastore$container_name,
    account_name = ws_blob_datastore$account_name,
    account_key = ws_blob_datastore$account_key,
    create_if_not_exists = TRUE)

datapath <- data_path(my_data, <path_on_my_datastore>)
dataset <- create_file_dataset_from_files(datapath)

See also