Log a predictions metric to a run

log_predictions_to_run() logs a metric score that can be used to compare the distributions of true target values to the distribution of predicted values for a regression task.

The predictions are binned and standard deviations are calculated for error bars on a line chart.

log_predictions_to_run(name, value, description = "", run = NULL)

Arguments

name
A string of the name of the metric.

value
A named list containing name, version, and data properties.

description
(Optional) A string of the metric description.

run
The Run object. If not specified, will default to the current run from the service context.

Examples

if (FALSE) {
data <- list("bin_averages" = c(0.25, 0.75),
             "bin_errors" = c(0.013, 0.042),
             "bin_counts" = c(56, 34),
             "bin_edges" = c(0.0, 0.5, 1.0))
predictions <- list("schema_type" = "predictions",
                    "schema_version" = "v1",
                    "data" = data)
log_predictions_to_run("mypredictions", predictions)
}