Upload files to the Azure storage a datastore points to

Upload the data from the local file system to the Azure storage that the datastore points to.

upload_files_to_datastore(datastore, files, relative_root = NULL,
  target_path = NULL, overwrite = FALSE, show_progress = TRUE)

Arguments

datastore
The AzureBlobDatastore or AzureFileDatastore object.

files
A list of strings of the absolute path to files to upload.

relative_root
A string of the base path from which is used to determine the path of the files in the Azure storage. For example, if we upload /path/to/file.txt, and we define the base path to be /path, when file.txt is uploaded to the blob storage or file share, it will have the path of /to/file.txt. If target_path is also given, then it will be used as the prefix for the derived path from above. The base path must be a common path of all of the files, otherwise an exception will be thrown.

target_path
A string of the location in the blob container or file share to upload the data to. Defaults to NULL, in which case the data is uploaded to the root.

overwrite
If TRUE, overwrites any existing data at target_path.

show_progress
If TRUE, show progress of upload in the console.

Value

The DataReference object for the target path uploaded.